requirements.txt
langchain
openai
tiktoken
deeplake
openai-whisper 
yt_dlp
streamlit 